+++ src/LogTable.php
@@ -4,18 +4,19 @@ declare(strict_types=1);
 
 namespace AchyutN\FilamentLogViewer;
 
-use AchyutN\FilamentLogViewer\Enums\LogLevel;
 use AchyutN\FilamentLogViewer\Filters\DateRangeFilter;
 use AchyutN\FilamentLogViewer\Model\Log;
 use Exception;
-use Filament\Facades\Filament;
+use Filament\Actions\Action;
+use Filament\Actions\ViewAction;
 use Filament\Infolists\Components\RepeatableEntry;
 use Filament\Infolists\Components\TextEntry;
 use Filament\Notifications\Notification;
 use Filament\Pages\Page;
-use Filament\Resources\Components\Tab;
+use Filament\Panel;
 use Filament\Support\Colors\Color;
-use Filament\Tables;
+use Filament\Support\Icons\Heroicon;
+use Filament\Tables\Columns\TextColumn;
 use Filament\Tables\Concerns\InteractsWithTable;
 use Filament\Tables\Contracts\HasTable;
 use Filament\Tables\Table;
@@ -29,12 +30,7 @@ final class LogTable extends Page implements HasTable
     #[Url(except: null)]
     public ?string $activeTab = null;
 
-    protected static string $view = 'filament-log-viewer::log-table';
-
-    /**
-     * @var array<string | int, Tab>
-     */
-    private array $cachedTabs;
+    protected string $view = 'filament-log-viewer::log-table';
 
     /** @throws Exception */
     public static function getNavigationLabel(): string
@@ -55,7 +51,7 @@ final class LogTable extends Page implements HasTable
     }
 
     /** @throws Exception */
-    public static function getSlug(): string
+    public static function getSlug(?Panel $panel = null): string
     {
         return self::getPlugin()->getNavigationUrl();
     }
@@ -87,9 +83,9 @@ final class LogTable extends Page implements HasTable
                 }
             })
             ->columns([
-                Tables\Columns\TextColumn::make('log_level')
+                TextColumn::make('log_level')
                     ->badge(),
-                Tables\Columns\TextColumn::make('env')
+                TextColumn::make('env')
                     ->label('Environment')
                     ->color(fn (string $state): array => match ($state) {
                         'local' => Color::Blue,
@@ -100,23 +96,23 @@ final class LogTable extends Page implements HasTable
                     })
                     ->toggleable(isToggledHiddenByDefault: true)
                     ->badge(),
-                Tables\Columns\TextColumn::make('file')
+                TextColumn::make('file')
                     ->label('File Name')
                     ->badge()
                     ->toggleable(isToggledHiddenByDefault: true),
-                Tables\Columns\TextColumn::make('message')
+                TextColumn::make('message')
                     ->searchable()
                     ->label('Summary')
                     ->wrap(),
-                Tables\Columns\TextColumn::make('date')
+                TextColumn::make('date')
                     ->label('Occurred')
                     ->since()
                     ->dateTimeTooltip()
                     ->sortable(),
             ])
-            ->actions([
-                Tables\Actions\ViewAction::make('view')
-                    ->infolist([
+            ->recordActions([
+                ViewAction::make('view')
+                    ->schema([
                         RepeatableEntry::make('stack')
                             ->hiddenLabel()
                             ->schema([
@@ -132,54 +128,18 @@ final class LogTable extends Page implements HasTable
             ->filters([
                 DateRangeFilter::make('date'),
             ])
+            ->deferFilters(false)
+            ->deferColumnManager(false)
             ->defaultSort('date', 'desc');
-    }
-
-    /**
-     * @return array<string | int, Tab>
-     */
-    public function getCachedTabs(): array
-    {
-        return $this->cachedTabs ??= $this->getTabs();
-    }
-
-    /** @return array<string, mixed> */
-    public function getTabs(): array
-    {
-        $all_logs = [
-            null => Tab::make('All Logs')
-                ->badge(fn () => Log::query()->count() ?: null),
-        ];
-
-        $tabs = collect(LogLevel::cases())
-            ->mapWithKeys(fn (LogLevel $level) => [
-                $level->value => Tab::make($level->getLabel())
-                    ->badge(
-                        fn () => Log::query()->where('log_level', $level)->count() ?: null
-                    )
-                    ->badgeColor($level->getColor()),
-            ])->toArray();
-
-        return array_merge($all_logs, $tabs);
-    }
-
-    public function getDefaultActiveTab(): null
-    {
-        return null;
-    }
-
-    public function updateTab(?LogLevel $level): void
-    {
-        $this->activeTab = $level?->value;
     }
 
     protected function getHeaderActions(): array
     {
         return [
-            \Filament\Actions\Action::make('clear')
+            Action::make('clear')
                 ->visible(Log::query()->count() > 0)
                 ->label('Clear Logs')
-                ->icon('heroicon-o-trash')
+                ->icon(Heroicon::Trash)
                 ->color(Color::Red)
                 ->requiresConfirmation()
                 ->action(function (): void {
@@ -195,8 +155,6 @@ final class LogTable extends Page implements HasTable
     /** @throws Exception */
     private static function getPlugin(): FilamentLogViewer
     {
-        $panel = Filament::getCurrentPanel();
-
-        return $panel?->getPlugin('filament-log-viewer');
+        return filament('filament-log-viewer');
     }
 }