+++ src/resources/views/log-table.blade.php
@@ -1,5 +1,3 @@
 <x-filament-panels::page>
-    <x-filament-panels::resources.tabs />
-
     {{ $this->table }}
 </x-filament-panels::page>